PHANGS pipeline creation.
