+++ Dockerfile
@@ -24,9 +24,9 @@ ARG OPENCADC_REPO=opencadc
 RUN git clone https://github.com/${OPENCADC_REPO}/caom2pipe.git && \
   pip install ./caom2pipe
   
-RUN git clone https://github.com/${OPENCADC_REPO}/blank2caom2.git && \
-  cp ./blank2caom2/scripts/config.yml / && \
-  cp ./blank2caom2/scripts/docker-entrypoint.sh / && \
-  pip install ./blank2caom2
+RUN git clone https://github.com/${OPENCADC_REPO}/phangs2caom2.git && \
+  cp ./phangs2caom2/scripts/config.yml / && \
+  cp ./phangs2caom2/scripts/docker-entrypoint.sh / && \
+  pip install ./phangs2caom2
 
 ENTRYPOINT ["/docker-entrypoint.sh"]
