+++ README.md
@@ -1,2 +1,2 @@
-# blank2caom2
-Starting point to build an application to generate CAOM2 Observations from FITS files.
+# phangs2caom2
+An application to generate CAOM2 Observations for [PHANGS](https://public.nrao.edu/news/2019-alma-phangs/) observations.
