+++ phangs2caom2/main_app.py
@@ -83,27 +83,29 @@ from caom2utils import ObsBlueprint, get_gen_proc_arg_parser, gen_proc
 from caom2pipe import manage_composable as mc
 
 
-__all__ = ['blank_main_app', 'update', 'BlankName', 'COLLECTION',
+__all__ = ['phangs_main_app', 'update', 'PHANGSName', 'COLLECTION',
            'APPLICATION', 'ARCHIVE', 'to_caom2']
 
 
-APPLICATION = 'blank2caom2'
-COLLECTION = 'blank'
-ARCHIVE = 'blank'
+APPLICATION = 'phangs2caom2'
+COLLECTION = 'PHANGS'
+ARCHIVE = 'PHANGS'
 
 
-class BlankName(mc.StorageName):
+class PHANGSName(mc.StorageName):
     """Naming rules:
     - support mixed-case file name storage, and mixed-case obs id values
     - support uncompressed files in storage
     """
 
-    BLANK_NAME_PATTERN = '*'
+    PHANGS_NAME_PATTERN = '*'
 
-    def __init__(self, obs_id=None, fname_on_disk=None, file_name=None):
+    def __init__(self, obs_id=None, fname_on_disk=None, file_name=None,
+                 entry=None):
         self.fname_in_ad = file_name
-        super(BlankName, self).__init__(
-            obs_id, COLLECTION, BlankName.BLANK_NAME_PATTERN, fname_on_disk)
+        super(PHANGSName, self).__init__(
+            obs_id, COLLECTION, PHANGSName.PHANGS_NAME_PATTERN,
+            fname_on_disk, entry=entry)
 
     def is_valid(self):
         return True
@@ -134,15 +136,14 @@ def update(observation, **kwargs):
     headers = kwargs.get('headers')
     fqn = kwargs.get('fqn')
     uri = kwargs.get('uri')
-    blank_name = None
+    phangs_name = None
     if uri is not None:
-        blank_name = BlankName(artifact_uri=uri)
+        phangs_name = PHANGSName(artifact_uri=uri)
     if fqn is not None:
-        blank_name = BlankName(file_name=os.path.basename(fqn))
-    if blank_name is None:
+        phangs_name = PHANGSName(file_name=os.path.basename(fqn))
+    if phangs_name is None:
         raise mc.CadcException(f'Need one of fqn or uri defined for '
                                f'{observation.observation_id}')
-
 
     logging.debug('Done update.')
     return observation
@@ -174,7 +175,7 @@ def _get_uris(args):
         for ii in args.local:
             file_id = mc.StorageName.remove_extensions(os.path.basename(ii))
             file_name = f'{file_id}.fits'
-            result.append(BlankName(file_name=file_name).file_uri)
+            result.append(PHANGSName(file_name=file_name).file_uri)
     elif args.lineage:
         for ii in args.lineage:
             result.append(ii.split('/', 1)[1])
@@ -195,7 +196,7 @@ def to_caom2():
     return result
            
 
-def blank_main_app():
+def phangs_main_app():
     args = get_gen_proc_arg_parser().parse_args()
     try:
         result = to_caom2()
