+++ scripts/config.yml
@@ -1,4 +1,4 @@
-working_directory: /usr/src/app/blankt2caom2
+working_directory: /usr/src/app
 # the netrc_filename value must be a fully-qualified name
 # netrc_filename: /usr/src/app/netrc
 # this is the name of the proxy certificate file supplied 
@@ -8,7 +8,7 @@ proxy_file_name: /usr/src/app/cadcproxy.pem
 # operational value is ivo://cadc.nrc.ca/ams
 # - modifies entries on sc2.canfar.net
 resource_id: ivo://cadc.nrc.ca/sc2repo
-# operational value is ivo://cadc.nrc.ca/ams/blank
+# operational value is ivo://cadc.nrc.ca/ams/shared
 # - queries entries on sc2.canfar.net
 tap_id: ivo://cadc.nrc.ca/sc2tap
 # file that contains the list of entries to process
@@ -47,8 +47,8 @@ progress_file_name: progress.txt
 # ad stream value - sorry
 # set stream to 'default', if using TaskType STORE or PULL
 stream: default
-archive: BLANK
-collection: BLANK
+archive: PHANGS
+collection: PHANGS
 #
 # CADC service execution metrics
 #
