+++ setup.cfg
@@ -22,13 +22,13 @@ omit = */tests/*
 minversion = 2.2
 norecursedirs = build docs/_build
 doctest_plus = enabled
-testpaths = blank2caom2
+testpaths = phangs2caom2
 
 [bdist_wheel]
 universal=1
 
 [metadata]
-package_name = blank2caom2
+package_name = phangs2caom2
 description = Starting point for application to generate CAOM2 observations.
 long_description = TBD 
 author = Canadian Astronomy Data Centre
@@ -36,12 +36,12 @@ author_email = cadc@nrc-cnrc.gc.ca
 license = AGPLv3
 url = TBD
 edit_on_github = False
-github_project = opencadc/blank2caom2
+github_project = opencadc/phangs2caom2
 install_requires = caom2utils caom2repo cadcdata
 # version should be PEP386 compatible (http://www.python.org/dev/peps/pep-0386)
 version = 0.1.0
 
 [entry_points]
-blank2caom2 = blank2caom2.main_app:blank_main_app
-blank_run = blank2caom2.composable:run
-blank_run_by_state = blank2caom2.composable:run_by_state
+phangs2caom2 = phnags2caom2.main_app:phangs_main_app
+phangs_run = phangs2caom2.composable:run
+phangs_run_by_state = phangs2caom2.composable:run_by_state
